+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Enhancements
 
 - UI improvements to the Sara status messages
+- Encourage Sara NOT to use the specific vector store filenames in her answers to the user - and instead use more general resource names like "Software Architectural Blueprint" - which are names also in the raw resources uploaded by Backend
 
 ### Bug Fixes